+++ Changelog.md
@@ -1,12 +1,27 @@
 # Changelog
 
+## [2.0.0] -2025-10-17
+### Added
+- Regrind Button
+- Moved admin button
+- Mva logo
+- Updated interface 
+
+### Changes
+- Dropdown box is now also searchable
+- Allowing editting in admin menu
+- Print preview matches print
+- Drop down menu automatically selects default printer
+- changed the font sans-serif typefaces for a cleaner look
+- Removed app name from applcation interface
+
 ## [1.1.0] - 2025-09-28
 ### Added
 - Scrap support for labels
 - Printer selection support (choose printer at runtime)
 - Editing in admin menu support
 
-### Fixed
+### Changes
 - Corrected output orientation from 4×6 to 8.5×11
 - Switched password storage from plain text to hashed values
 - Changed Print (ctrl-p) to just print